// client/src/app/services/polygon/polygon-enum.ts
export enum PolygoneConstants {
  MAX_VALUE = 12,
  MIN_VALUE = 3,
  SIDE = 3,
  ANGLE_PI = 180,
  SIDE_VALUE = 60,
  PLACEMENT = 0,
  SIDE_MULTIPLE_3 = 3,
  SIDE_MULTIPLE_4 = 4,
  SIDE_MULTIPLE_5 = 5 ,
  DEGREE_MULTIPLE_3 = 78,
  DEGREE_MULTIPLE_4 = 55,
  DEGREE_MULTIPLE_5 = 60,
  DEGREE_MULTIPLE_7 = 37
}